// src/generated/models/Dev_table_facultiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Dev_table_facultiesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Dev_table_facultiesstatecode = keyof typeof Dev_table_facultiesstatecode;
export const Dev_table_facultiesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Dev_table_facultiesstatuscode = keyof typeof Dev_table_facultiesstatuscode;

export interface Dev_table_facultiesBase {
  dev_namefaculty?: string;
  dev_table_facultyid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Dev_table_facultiesstatecode;
  statuscode?: Dev_table_facultiesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Dev_table_faculties extends Dev_table_facultiesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}